TC_16.001.01 moved methods to ManageJenkinsTest
